Promote _gasneti_mythread_slow and friends to top level
This is mostly just a MOVE of existing code from gasnet_extended_help.h
to gasnet_help.h. This change is desirable for its own merits, but
is now required to break a cycle on the FARG goop for gasneti_AMPoll
in gasnet_help.h in the EX branch.

The only content changes:

* Stat event for DYNAMIC_THREADLOOKUP is factored out to avoid a
  dependency cycle with gasnet_trace.h

* _GASNETE_MYTHREAD guard macro is renamed to _GASNETI_MYTHREAD_SLOW