aries/gemini: reduce notify ring and Cq sizes
This commit corrects over-allocation of both the notify ring and the
Cq used for incoming AM traffic.

The notify ring was oversized in the (non-default) case that the
NETWORKDEPTH_TOTAL (bounding the number of outstanding Requests) was
less than the "credits" (which don't count messages in the traditional
manner in the Eager protocol).

The Cq was oversized by nearly a factor of two due to again failing to
properly consider NETWORKDEPTH_TOTAL as distinct from the max credits.
In addition over-estimating the number of Requests awaiting a Reply,
the number of Cq entries required for incoming Replies should *not* be
scaled with remote process count.