aries/gemini: Convert ctrl/credit to CqWrite
This commit converts shutdown and credit-return messages into CqWrite
operations.  The credit-return message has become a control message
(like the shutdown) rather than a special variant of a Reply.  The
control message path no longer uses the notify ring, and instead sends
all of its data via a CqWrite operation.

The lack of notify ring use for the shutdown message means the nominal
size of the ring is reduced by one.  However, since the ring size is
rounded up to a power-of-two, this either leaves the actual size
unchanged or cuts the size in half!  For the defaults the size has not
changed.

While this commit does not reduce the size of the notify ring for the
common case, it does lead to a measurable reduction in the latency of
the credit-return message sent when a Request handler does not send an
explicit Reply.  This results in measurable performance gains on
testam's one-way flood tests.

This commit is also the first instance of mixing CqWrite and "normal"
target notification events in the same Cq.  Further use of this mixing
is expected in the future.