AM: eliminate thread lookup for NBRHD PrepareReply
This commit restructures the code for the reference implementation of
NP-AM's gasnetc_AM_PrepareReply{Medium,Long}() to swap the relative
order of the in-nbrhd check and the sd-lookup in thread-specific data.
This eliminates the need for a dynamic thread lookup for NP-AM Replies
which are in-nbrhd.

The clone of the reference to template-conduit and gemini-conduit are
updated as well.  In the case of gemini-conduit, this also corrects
incorrect code for the in-nbrhd case.