Fix GASNet AM handler dispatch to be strictly conforming
GASNet previously assumed handler functions could be safely dispatched
by casting the handler function pointer to a varargs function pointer
type and invoking the result with the correct number of arguments.
However, this is not guaranteed to work by C language rules (because the
user's AM handler function is not declared using varargs, hence the
function types at definition and invocation are incompatible), and we've
now seen some evidence (bug 3679) that it can break for compilers and
architectures we care about.

This commit changes AM handler dispatch logic to cast the handler
function pointer to a function pointer type with exactly the correct
signature (and fixed number of arguments).