aries/gemini: improve GASNET_NETWORKDEPTH handling
This commit improves the behavior for GASNET_NETWORKDEPTH larger than
GASNET_NETWORKDEPTH_TOTAL when the Rendezvous protocol is used, in
response to PR feedback.

The value of the former was already reduced to never exceed the later,
but too late to effect the value of the credit counter.  This could
not have lead to incorrect behavior.

The main change is that we now warn when reducing the user's request.