tests: correctness testing of ReduceToAllNB
This commit adds correctness testing of gex_Coll_ReduceToAllNB() to
the existing testreduce.c.  Specifically, a random 20% of the
reductions are now reduce-to-all.

Additionally, this commit adds "scribbling" over the output arrays
prior to each reduction call.  This should eliminate the possibility
that the dynamically allocated output array already contained the
expected values.