coll-ex: SUBORDINATE support in Bcast and Reduce
This commit adds support for "SUBORDINATE" calls to the GEX interfaces
for Broadcast and ReduceToOne.  This support was omitted in previous
work to lessen the effort, given that there was no need at the time.
Now this support is needed for the (trivial) initial implementation of
ReduceToAll.

This commit introduces an internal GASNETI_FLAG_COLL_SUBORDINATE flag
for use in GEX interfaces, which mirrors (same value, though that is
not enforced) the legacy GASNETE_COLL_SUBORDINATE flag.  This is done
mainly to avoid conflating the values for the legacy collective flags
with the value of gex_Flags_t.  However, this should be revisited at
some point in the future when the GEX collectives out number the
legacy ones, and in particular when segment disposition flags are
defined and honored.  At that point, it will likely make sense to
define the legacy collective flags as (deprecated) aliases for GEX
flags.