coll-ex: fix progress issue with ReduceToOne
This commit adds a missing AMPoll() at the end of the implementation
of gex_Coll_ReduceToOneNB(), as is already present for all other NB
Collective initiation functions (except Barrier, which is distinct).

The issue is that collective initiation on its own only queues work
for the progress engine, but does not itself run that engine.  Thus
for a long stream of back-to-back initiations (as is exercised in both
testreduce and testcollperf's "reduce_nb") the list of active
collectives keeps growing without actually attempting *any*
communication!  As a result the times reported by testcollperf for
"reduce_nb" were significantly longer than for the blocking "reduce"
test, and the discrepancy would grow with iteration count.