aries/gemini: make more explicit segment queries
This commit replaces calls to gasneti_in_segment() with "fullsegment"
or "clientsegment" versions.

In the case of the Put and Get paths, we previously checked both
in_segment and in_auxsegment before resorting to dynamic registration.
However, since in_segment is equavelent to in_fullsegment in internal
code, the auxseg check was redundant.  We now just make the fullsegment
query.

In the case of one error reporting path, the equivalent of in_segment
and fullsegment would have mistakenly reported local addresses in the
aux seg as if they were in the client segment.