aries/gemini: fix a fence post error
This commit fixes an erroneous call to gasneti_in_auxsegment() with
_nbytes=0, which would return true for one byte past the end of the
aux segment.  This became a problem recently, when rounding up the aux
seg size to hugepage size resulted in the aux and client segments
being adjacent.  This caused the conduit to issue RMA with an address
of the first byte of the client segment using the memory handle for
the *aux* segment!

Additionally, this commit adds an assertion in gasnet_help.h to
disallow passing _nbytes == 0 to the in-segment and bounds checks,
and fixes a typo in their comments.