Remove all FCA support from GASNet-EX
This commit removes all FCA support from GASNet-EX. We no longer have
an active maintainer for FCA, and our current code appears likely to
be incompatible w/ upstream changes (at a minimum configure does not
find the new "hcoll" naming).

So, we are removing the current FCA code with hopes that we can
reintroduce it when the EX collectives (re)implementation has advanced
further.