coll: eliminate gasnete_coll_op_info_t
This commit eliminates gasnete_coll_op_info_t, absorbing it entirely
into gasnete_coll_scratch_req_t.  The lifetime of the former was
already a subset of that of the later, and contained only one data
field (op) not already present in the scratch_req or the op.  So,
only the linkage and op fields are added to scratch_req.