coll: reduce malloc/free in scratch management
This commit further reduces the use of dynamic malloc/free in the
scratch space management logic.

In the case of `config->peers`, the allocation is inlined into the
scratch_config type for the case of `num_in_peers <= 8`.

In the case of `op->scratchpos`, the allocation is unconditionally
merged with that of `scratch_req->out_sizes`, for which dynamic
allocation has already been eliminated for out-degree no larger than
eight.  Now these two share space such that there is at most one
dynamic allocation, and only for the cases of tree ops with out degree
larger than eight or disseminations ops with more than 15 phases.

These two geometry-dependent allocations are the only dynamic
allocations related to scratch space management.  All others use free
lists to avoid dynamic allocation in the common case.