RAtomic: fix validation on sub-teams (Bug 3783)
This commit repairs the flawed domain validation logic in
gex_AD_Create() reported in Bug 3783.  Specifically, the check for
single-valued arguments was always performing its collectives on team0
instead of the passed 'tm'.

In addition to simply using the right team, this logic is also updated
to use reductions, instead of potentially memory-intensive Exchange.