coll-ex: fix a goto-bypassed-init warning
This commit's main purpose is to relocate the closing brace of a block
which exists to permit declaration of a variable local to a one "case"
of a switch.  The previous "late" placement meant that a goto jumped
into the block bypassing the variable initialization.

The bulk of this commit is tangentially-related improvements to the
four similar switch statements in this file to improve readability
and optimization.