Add an internal _tm0 field to gex_Client_t
This commit adds a _tm0 field to Client, for internal use.

The new field is initialized in gasneti_tm_alloc() which is modified
to use NULL-valued _tm0 to detect the first call per Client, rather
then depending on a boolean argument passed by the caller.  Removing
that argument accounts for the bulk of this commit.

The types Client <- EP <- TM <- Client now create a cycle, meaning we
can no longer use anonymous structs for all three types.  Therefore,
this commit introduces a named struct type for TM.