tests: change default source memory to in-segment
The following tests:

* testsmall
* testlarge
* testqueue
* testalign

now all default to using in-segment source memory. With this change,
all GASNet tests with optional segment disposition now default to
in-segment, which is usually expected to deliver the best performance.

In all cases, the old behavior can be restored by passing the `-out`
command-line option to request out-of-segment source memory.