fixup: Add an internal _tm0 field to gex_Client_t
This commit reworks the cycle-breaking use of a tagged structure for
gasneti_TM_t to avoid running afoul of the C++ type rules.

In C, the internal TM structure is tagged and the _tm0 field in Client
is fully typed and thus C code using it is fully type-checked.
However, in C++ the structure is anonymous and the _tm0 field in
Client is a pointer to incomplete type.  This means that any eventual
use of this field in C++ code will require a cast.