ibv: implicit ODP replacement for local firehose
This commit uses "implict ODP" (the registration of the entire address
space) as a replacement (when available) for local-only firehose in a
SEGMENT_FAST configuration.

If ODP is disabled or not found at configure time, the code is
omitted.

If ODP is enabled at configure time, but the necessary "capabilities"
are not found at runtime, a warning is printed and execution proceeds
without ODP.