testslice: support segment larger than INT_MAX
This commit updates testslice to use size_t for the segment size,
rather than int, to allow for sizes larger than INT_MAX.

Additionally, the random initialization of the source data is changed
from 1-byte to 8-byte granules.  This takes advantage of the 64-bit
TEST_RAND() and noticeably speeds startup.