pmi-spawner: detect non-single-valued 'len'
This commit adds assertions in pmi-spawner to detect over- and
under-runs in the encoding of the transferred data.  Such conditions
could potentially be caused by data corruption, but are certain to
occur if the 'len' arguments to a bootstrap collective are not
single-valued.

(cherry picked from commit 20309d6710ab94878f071fd8bcdad6880deb8c4c)