coll-ex: RVous Broadcast via binomial tree
This commit updates the logic for RVous-based Broadcast (used for
out-of-segment src and dst when too large for Eager) to use a binomial
tree, rather than the previous "flat tree".  This may not be the
optimal geometry, but should be a measurable improvement over the
previous implementation.

Similar changes for Scatter and Gather are possible, but are
non-trivial because use of some sort of scratch space would be
required at the internal nodes of the tree.