NP-AM: ensure alignment of gasnet-owned buffers
This commit adds an assertion in the NP-AM Prepare path to ensure that
gasnet-owned buffers are always 8-byte aligned.  Additionally the one
known alignment violation (PSHM NP-AM Long) is corrected.

The lack of alignment in PSHM/Prepare*Long came to light due to SIGBUS
on SPARC ILP32 in gasneti_init_sd_poison().  However, the need for
alignment is not limited to that debug-only function - it is needed to
ensure the client can safely assemble their payload into the buffer.