Fix bug 3797: invalid asm on Solaris/SPARC
This commit resolves bug 3797 (gcc inline asm for SPARC v8plus
incorrect on Solaris) by excluding FP regsters from consideration when
chosing a register pair for the `ldd` or `std` instructions.

This seems to be a matter of the GNU vs Solaris assembler, not Linux
vs Solaris ABI.  While `gas` converts an `LDD` with FP register to
`LDDF`, the Solaris assembler does not.