ibv: warn if ibv-conduit is used on OPA hardware
This commit adds logic to warn if ibv-conduit used for the "hfi1_*"
HCA name used by Intel Omni-Path.  This recommendation (also in the
ibv-conduit README) is based on the observation (in bug 3339) that
ibv-conduit significantly under-performs on OPA hardware (though the
difference appears to have lessened since those results were
measured).