Fix gasnett_backtrace_user
Prior work on bugs 2117 and 2554 left this particular tentative
definition in a "grey area" of the C spec. Specifically, the
tentative definition provided by gasnet_tools.c should not have a
preceding `extern` declaration, as the spec is unclear whether that is
required to work.

This issue was not known to cause any problems in practice, but it
seems best to use the same convention for declaring all of the
library's tentatively defined symbols.