coll-ex: Update GEX barrier for improved semantics
This commit updates the implementation of gex_Coll_BarrierNB() (and
the undocumented blocking variant) to utilize the existing "consensus
barrier" logic in the collectives, instead of using the GASNet-1
barrier implementation directly.

This change means that GEX barriers can be interleaved with
collectives (including other GEX barrier calls) in any order that
remains collective.  This includes having multiple GEX barriers
outstanding.  This change is necessary (and sufficient) to allow the
UPC++ barrier to be implemented directly in terms of GEX's barrier.

The "consensus barrier logic" exists to serialize access to the one
GASNet-1 barrier implementation by the (potentially numerous) barriers
used internally within collectives to implement synchronization as
required by the sync flags.  So, use to serialize multiple instances
of the GEX barrier is a natural fit.

This implementation is over-synchronized in two ways.  The first is
that no communication on a GEX barrier can begin until the proceeding
barrier is locally/internally completed.   This means a potential
increase in latency relative to an implementation which could
"pipeline".  The second is that this ordering includes barriers
internal to collectives.  However, I do not expect to see significant
mixing of the GEX barrier with IN_ALL or OUT_ALL collectives (at least
not until the GEX barrier implementation is revised at least once
more).

Future work on separating the GASNet-EX barrier implementation from
the GASNet-1 version (with its UPC-style name matching) may supersede
this work.  However, in the meantime this provides the desired
semantics while retaining use of conduit-specific barrier
implementations where applicable.