tests: fix non-collective behavior in testimm
This commit corrects non-collective behavior in testimm, in which
ranks other than 0 would incorrectly execute barriers corresponding to
their "passive" role in Put and Get tests which were not actually run.

This should correct TIME failures in nightly runs of this tests, which
had previously been ignored as due to slow testers.