Chapel/Legion CI: preserve sub-make exit codes
This commit replaces multiple instances of `cmd | tee logfile` in the
Chapel and Legion CI Makefiles with portable Bourne shell code to
return the exit codes from `cmd` as the result of the compound
command.  Previously, the successful return from `tee` was masking any
failure of the `cmd`.

The `cmd` is a sub-make in all current cases, but the logic does not
depend on that in any way.

(cherry picked from commit e9a0d95550618dac21255acf7bf06b2dd033019c)