gemini/aries: tolerate unequal GASNET_DOMAIN_COUNT
This commit addresses hangs reported as chapel github issue 11621 by
ensuring that the auxseg size requests will be single-valued, even
when the domain counts are not.  This is accomplished by using the MAX
of domain counts for auxseg (pinned memory) requests, while continuing
to use the local domain count for allocation of GNI resources.