Debug GASNet mutexes: Clean up ownership handling
In GASNET_DEBUG mode, gasneti_mutex_t contains additional state to
track the current owning thread, to detect usage violations.
This state was making assumptions about pthread_t not guaranteed by
the POSIX spec - most notably that it is a scalar type.

This commit removes that assumption, replacing equality comparisons
on pthread_t with pthread_equal() as recommended by POSIX.
POSIX unfortunately does not provide an "invalid value" for pthread_t so
we still supply our own - but the value is chosen to have less chance
of collsion with a valid thread id, and is now easy to replace at compile
time if necessary.

Also add some macros to improve encapsulation of ownership information.

Adjust the few conduit places that violate the abstraction accordingly.