ibv: rearrange/streamline gasnete_put_nb()
This rearranges the code in gasnete_put_nb() to fix a persistent
(harmless but annoying) warning from gcc-7 (and others).

IMHO the result is cleaner, and should hopefully improve NDEBUG codegen
by reducing branches and by removing or tightening the scope of temporaries
that the analysis might not adequately understand (as evidenced by the
false maybe-uninit warnings).

It should be semantically identical, except that it won't detect/diagnose
invalid lc_opt inputs in NDEBUG mode (saving one branch).