ibv: silence a harmless warning from gcc-7.3.0/aarch64
Insert a dead store along two control paths in gasnetc_AMReplyLong()
to avoid the maybe-uninitialized warning from gcc-7.3.0/aarch64.

This is a bit of a hack. A better solution would be to rearrange
the gasnetc_AMReplyLong() function (and other AM injection functions
in this file) along the manner of the previous commit. However, in this
case the common call being factored across LC mechanisms (gasnetc_ReqRepGeneric())
is a very large function marked as GASNETI_INLINE (and probably should
not be, see bug 1879) so cloning the call would induce a large code
expansion for gasnetc_AMReplyLong() that might negatively impact performance.