atomics: fix mismatch behavior with big-endian XLC
This commit corrects the logic for inline asm with "mismatched"
IBM XL compliers (ones not probed by configure).  Previously I had
assumed that GCC style inline asm was available in all XLC versions
12.0 and newer.  While that is true as far as "available" goes, I
had overlooked the fact that support in the big-endian XLC compilers
is not sufficient to support use of GCC style inline ASM for the Tools
atomics.