Fix bug 3812 - intermittent ENOMEM at startup
This commit resolves
  bug3812: Intermittent mmap failures at startup: cannot allocate memory

This commit adds logic to gasneti_mmap_segment_search() so that a
failure of the final fixed mmap() is not necessarily fatal.  If the
mmap fails with ENOMEM, then a bounded linear-descending mmap probe is
run starting at the just-failed size.  If that also fails, the entire
probe is restarted (at most once) with maxsz set to the size at which
the linear-descending probe failed.

Additionally, a long-latent bug for BG/Q+POSIX and Cygwin+SysV (both
non-default) is corrected by taking care not to discard the return
value from gasneti_do_mmap_fixed when GASNETI_PSHM_MAP_FIXED_IGNORED
is defined.