ibv: strengthen ODP configure chsck
This commit modifies the ODP configure check to more closely mimic the
actual usage.  In particular, all struct members and constants used in
the code should now appear in the probe.

This resolves a problem observed with an older OFED-3.3 installation
in which the configure probe determined ODP was present, but the ODP
code in ibv-conduit could not compile.