Update Cray cross-configure scripts
This commit adds the CROSS_CACHE_LINE_BYTES setting to the four
current cross-configure scripts for Cray systems.  This eliminates an
unintended difference from the current template.

This has the side-effect of brining the default cache line size for
these systems into agreement with what the vast majority of our
testing has used.  There should also be minor memory savings due to
reduced padding in, and alignment of, certain structures.