testtools: Reduce false negatives from gasnett_tick_overheadus test
testtools previously compared the reported overhead and granularity of
the tick timer and reported an error when they looked inconsistent.
However the algorithm used to measure these quantities is susceptible
to reporting falsely high overhead in the presence of high system
noise and/or a poorly timed context switch during the measurement,
and the ensuing error led to a harness false negative that is almost
always harmless (increasing tester noise and potentially obscuring
other, more serious failures).

Downgrade the failure in question to a "WARNING" so that message alone
will not trigger a test failure.