ibv: use AC_FUNC_ALLOCA outputs
This commit updates the conditional use of alloca() by ibv-conduit to
properly use the outputs from the AC_FUNC_ALLOCA() macro.  We now (a)
disable use of alloc() if the probe did not find support, and (b) do
not include alloca.h unless the probe found that it exists.

The second change is necessary to compile ibv-conduit on FreeBSD-12.0
where alloca() is prototyped in stdlib.h, and there is no alloc.h.