ssh-spawner: correct strncpy() usage
This commit corrects the use of `strncpy()` for the SSH_REMOTE_PATH
value to ensure a '\0' terminator is always present.  Additionally,
the allocated space now honors `PATH_MAX` if defined (with an
unchanged 1024-byte default if it is not).

(cherry picked from commit 1eba41ede508f221bbeeac9ec29cda5d172e01bf)