gasneti_console_messageVA: reduce output interleaving for short messages
Use a short string buffer on the stack to attempt complete message formatting.
When it succeeds (ie the message is sufficiently short) this allows
us to send the formatted buffer to stderr in a single fputs operation,
which has been observed to reduce sub-line output interleaving for processes
whose output is not being directed through a line-buffering layer -
notably including smp-conduit.

By default, all messages with a prefix under about 128 characters and total
length under 256 characters qualify as "very short".