Tools: remove unused GASNETC_SET_AFFINITY hook
This commit removes the GASNETC_SET_AFFINITY() hook intended for
conduit-specific implementation of gasnett_set_affinity().

This hook is currently unused and would *not* function as intended if
a conduit did define this identifier, because gasnet_tools.c does not
include *_fwd.h (by design).  If we ever need this override, it should
be re-engineered to use a function pointer as with the
gasnett_getenv() override.