Tools: update gasnett_set_affinity()
This commit adds an integer return type (0 on success) to
`gasnett_set_affinity()` and a `GASNETT_SET_AFFINITY_SUPPORT` define
to indicate if GASNet Tools has support for the current platform.
This includes updates to the docs, implementation and tests.

Ironically, however, we also update the docs to mark this interface as
deprecated (suggesting use of hwloc).