gasnet-tests CI: Add g2ex annotations
The g2ex feature is used to blacklist G-1 tests that rely upon
G-1-specific APIs that are not supported in g2ex (mainly, G-1 collectives),
so this suite can be run against GASNet-EX to validate G-1 compliance.

This is intentionally a G-1 only commit, as it's not relevant to the EX
harness.