pshm: prefault PSHM memory region
This commit ensures that at least one byte is written to every page of
the PSHM memory region (allocated by gasneti_pshm_init()) prior to
reaching gasneti_mmapLimit().  This ensures (a) that a later
gasneti_mmapLimit() probe will be reduced by the pages needed for PSHM
internals and (b) that later use by PSHM internals will not experience
a fatal SIGSEGV or SIGBUS due to an inability to allocate a page
frame.  Additionally, this change should result in predictable
affinity (though it looks likely that the majority of the memory
either has no natural "owner" or was already certain to receive first
touch by the appropriate owner).

This commit provides a new gasneti_pshm_prefault(addr,len) to assist
in this task.

Within gasneti_pshm_init(), the shared "info" portion is prefaulted by
"rank0", and the buffers used to construct out-bound AMs are
prefaulted by their respective owners.  The portion containing message
queue heads and tails have always received complete initialization.

Currently, only smp, gemini/aries and ibv allocate space for conduit
use in the PSHM memory region.  In the cases of smp and gemini/aries
this memory was already initialized immediately after allocation.

In the case of ibv-conduit, a portion of the space was already
initialized immediately, but the part for XRC coordination was not.
So, this commit includes new logic to ensure that "rank0" calls
gasneti_pshm_prefault() on the XRC table of RCV QPNs, and the XRC init
logic is reordered slightly to move a pre-existing barrier after the
prefaulting.

Additionally, the template comment regarding gasneti_pshm_init() is
updated, including in two conduits which have preserved this comment.

This commit addresses ibv-conduit's Bug 3693: "Unexpected error Bad
address (errno=14) when registering the segment" by ensuring that the
statvfs() call in gasneti_mmapLimit() will correctly count space
consumed by the PSHM memory region (for at least POSIX-based PSHM on
Linux).