coll: select large starting sequence numbers
This commit changes the starting sequence numbers for collectives and
consensus barrier to be 0xffff,fff8, so that wrap-around will occur
relatively early in runs of tests with non-trivial use of collectives.
This is intended to improve debugging coverage, but is not dependent
on GASNET_DEBUG since the starting points are entirely arbitrary.