bug 3866: fix a potential overflow in the lo calculation
It's conceivable that a rare set of circumstances (even in the
absence of tick monotonicity violations) could lead to an integer overflow
in the calculation of lo that might result in a corrupted calibration
matching the behavior reported in the bug.

Take steps to help encourage such overflows to instead be discarded.