Revert "ibv: correct too-large MaxLocalSegmentSize"
This has been determined that the subtraction of the aux segment size
from the result of the mmap limit probe is *only* correct if the probe
result was constrained by the availability of shared memory (for
instance the size of /dev/shm).  However, if the limit is due to a
user's environment settings, then it is *incorrect* to reduce the size.

This reverts commit bb432683cde2dc4d95bbb6819d336dbc946cea43.
This reopens Bug 3693:
  Unexpected error Bad address (errno=14) when registering the segment