Implement gasneti_sharedLimit() to fix bug 3874
This commit implements and applies gasneti_sharedLimit() as requested
in bug 3874.  This separates the concerns of maximum mmap()able space
from the maximum space that can be shared (for instance due to space
on a filesystem).

Note that while the original problem (bug 3693) was specific to ibv,
*all* conduits are updated to use the new gasneti_sharedLimit() in
determining the arguments to gasneti_segmentLimit().  This is based on
the observation that the FILE and POSIX cases (the only ones which
impose any limits currently) are limits that exist irrespective of any
swap space which may exist.  So, while bug 3693 showed an "early"
failure due to memory registration, an application on any conduit
could experience a SIGBUS if/when lazy allocation of page frames
eventually exceeds the capacity of the backing-store.  This has been
confirmed in actual testing with udp-conduit.