Bug 3876: Tweak PLATFORM version encoding for Intel
The current allocation of bits in PLATFORM_COMPILER_VERSION for Intel to accommodate
the oddball patch convention also mean that the 21.0.0 release of icc expected
in two years will generate a PLATFORM_COMPILER_VERSION value that overflows a
32-bit signed int. We don't document hard guarantees about the range of this
value, but for most (all?) compilers we've tried to keep the encoding under 32
bits to help prevent accidents.

Reducing the Maj/Min shift from 20 to 19 raises the high-side overflow in the
encoding from release date 2021 to 2041.

Raise the _PLATFORM_COMPILER_INTEL_MIN_BUILDDATE (the base value used for
offset encoding) from 1970 to 1990, which raises the low-side overflow in the
encoding from release date 2023 to 2043.

Hopefully before 2041 the Intel compiler will straighten out its
version-reporting mess so we can switch to something more user-friendly.