test_internal: don't leak NBI comms from pf test
This commit reworks the communication in the progress functions test
to use a mixture of NB and NBI access regions.  The events are saved,
and are synced either by a later run of the progress function, or by
the main code after stopping the progress functions.  This avoids
"injecting" of NBI comms into the iop/aop of the context running
outside of the progress function.