aries: feature macro for Aries CE support
This commit adds a GASNETC_BUILD_GNICE macro (undefined or 1) to
control (yet to be implemented) build of Aries Collective Engine (CE)
support.

Additionally, GASNETC_LOCAL_GNI_EP is added as the logical OR of the
new GASNETC_BUILD_GNICE and the pre-existing GASNETC_BUILD_GNIRATOMIC.
It replaces instances of the latter where used to allow/prevent
GNI-level endpoint connecting processes on the same host.