aries: refine gasnetc_post_ce, add gasnete_test_ce
This commit refines the implementation of gasnetc_post_ce(), moving
the output struct (which must reside in registered memory) out of the
post descriptor and into its own portion of the aux seg.  This change
eliminates the need for a "long lived" post descriptor.

Additionally, this commit adds gasnete_test_ce() to contain the
updated logic for completion of a CE operation.

Finally, the validation/example code in gasnetc_init_ce() is updated
accordingly.