coll-ex: override for consensus barrier "guts"
This commit provides the means for a conduit to override the consensus
barrier "guts", which consist of a simple (no UPC-style name matching)
split-phase barrier.  This is an finer-grained alternative to the
(still available) override that allows a conduit to complete replace
the consensus barrier "component".

In the absence of a conduit-provided override, this commit makes no
actual change to the consensus barrier logic.