aries: hook consensus barrier to use Aries CE
This commit takes advantage of the recently-added fine-grained
consensus barrier overrides to plug in a simple split-phase barrier
using the Aries Collectives Engine (when available).  When the
Aries CE facility is unavailable, the implementation remains
unchanged other than an extra branch.

Since the consensus barrier is the basis of gex_Coll_BarrierNB(), this
yields measurable improvement in barrier performance (in the
neighborhood of 40% in cursory testing on Cori Haswell nodes).