ofi: implement provider selection filter
This commit adds provider filtering to ofi-conduit.

If a provider was selected at configure time (by the user or by
probing `fi_info`) then *only* that provider will be used.  This
seems important to ensuring safe use of the settings of
  GASNETC_OFI_HAS_MR_SCALABLE
  GASNETC_OFI_USE_THREAD_DOMAIN

If no provider was selected at configure time, then the entire
prioritized list of providers maintained in configure.in is
used by the conduit at run time to select the most-preferred
of the supported conduits (and thus we use an unknown provider
if and only if no supported conduit can be found).

The behavior for "Invalid or unsupported OFI provider specified via
the --with-ofi-provider flag" remains unchanged.  So,
`--with-ofi-provider=can-and-string` or `FI_PROVIDER=none` will
*continue* to defer provider selection until runtime.

This filtering has the ironic side effect of rendering the recent work
to support 'tcp;ofi_rxm' and 'verbs;ofi_rxm' providers unreachable,
since 'sockets' provider is given higher priority.

(cherry picked from commit 7b5114e73451312e1b498620ab193b8fd73a9e3c)