aries: add PSHM-hybridization to CE barrier
This commit add logic to the Aries CE barrier implementation to allow
for more than 30 processes per host, so long as they are grouped into
no more than 30 neighborhoods (aka supernodes).

The use of process is still preferred over neighborhoods when possible
because the latter requires funnelling the CE communications through a
single (statically-selected) process per neighborhood.  This is a
consequence of the static configuration of the CE tree topology.

Initial timing between a pair of hosts on Cori Haswell nodes show the
hybrid to be faster in all cases where the previous implementation
would have used the GNI dissemination barrier between the
neighborhoods (and the more general PSHM barrier within each
neighborhood).