ibv: add IBV_ACCESS_REMOTE_ATOMIC flag as needed
This commit adds IBV_ACCESS_REMOTE_ATOMIC to the access flags passed
to segment memory registration and QP creation.  Some factorization
is also applied.

This does not add remote atomic access rights to bounce buffers or AM
receive buffers, though that could be done in the future with little
work.