ibv: Implement and document GASNET_USE_FENCED_PUTS
This commit implements insertion of an Atomic after each Put on
multi-rail systems, using the CQ entry for the Atomic as the
indication of remote completion of the Put.  This is required for
strict mapping of GASNet's semantics onto InfiniBand, as described in
"Bug 3447 - ibv-conduit ordering problems with multiple HCAs".

As documented in ibv-conduit's README, a new environment variable
`GASNET_USE_FENCED_PUTS` is a Boolean to control this behavior, and
defaults to OFF.  If enabled when only a single rail is used, a
warning is issued.  The text of the warning differs depending on
whether multi-rail was enabled at configure-time.

In a default `--disable-ibv-multirail` build the support code is not
compiled.