ibv: append helpful text to XRC init fatal errors
This commit addresses the concern raised in bug 3937 regarding how a
failure to initialize XRC support may result in an error message that
the end-user cannot understand.  The only changes are to append a
suggested work-around to several error messages which are believed
to occur only when XRC support is missing or badly broken.