harness: repair harness.conf for recent changes
This commit updates the harness.conf for the gasnet-tests suite to
accommodate the recent changes to several tests' output (to use
GASNet-EX names).  This includes raising the OutputLimit for
testsmall and changes to several BenchmarkResult patterns

Some of the BenchmarkResult patterns have apparently been broken since
43c95c34 in June 2016, where '_bulk' was stripped from several
instances of 'get_nb_bulk'.

Additionally:

+ The AppArgs for testlogGP-par have been incorrect since some time in
  2006 when the pollcnt argument was added, resulting in an unmatched
  BenchmarkResult (as well as different runs for par-vs-seq).
+ We have been collecting blocking Put latency from testval for both
  seq and par.  This commit starts collecting Get latency from the par
  run, consistent with several other tests