ibv: remote atomic results handling
This commit adds the necessary logic and storage to manage the result
location for offloaded remote atomics in ibv-conduit.  All atomics are
fetching.  So for the non-fetching case we need AuxSeg space as a
"sink" for the result.  For the fetching case this commit adds both
zero-copy and bounce-buffer atomics cases to the progress engine
(completion handling).  The latter is required because the result can
only be written to registered memory.

The AuxSeg use is unconditional, but is shared with the
conditionally-compiled "Fenced Puts" support.  So, a portion of this
commit is rearrangement of that logic.  Note, however, that the new
(conditional) logic stores only a local address, not an O(N) table of
remote ones.

There is not yet any logic to actually initiate offloaded atomics.