Further improve core binding support with Open MPI
This commit resolves the known conflict between `-N` and `-c` options
to `gasnetrun_mpi` when backed by Open MPI 2.x or 3.x.  Additionally,
the support for core binding is extended back to Open MPI 1.8.

For Open MPI 1.[567], the support for `-N` remains unchanged, but the
implementation of `-c 0` is corrected to use `--bind-to-none` rather
than `--bind-to none` (which is not supported prior to 1.8).

Testing:

  + Open MPI 1.5.4  on Solaris nodes of Dirac
  + Open MPI 1.8.4  on SDSC's Comet
  + Open MPI 1.10.0 on Dirac
  + Open MPI 2.1.2  on Dirac
  + Open MPI 3.0.0  on Dirac
  + Open MPI 3.1.3  on OLCF's Wombat
  + Open MPI 4.0.0  on Dirac