Improved handling of generic and O/S atomics
This commit leverages the configure-time availability of the atomics
implementation to ensure configure-time disable of PSHM and
ibv-conduit if using generic or OS atomics.

Previously the PSHM configure logic was sufficient only to notice when
configured with --enable-force-{generic,os}-atomicops.  If the
platform was using (not-forced) generic or O/S atomics, then a failure
would occur at build time instead.  For ibv-conduit there was no
configure-time check and the failure could only appear at build time.

This new logic will also reject "hybrid" as the primary atomics
implementation.  However, this should only ever be used for the
64-bit fixed-width atomics.